intents: the last two places a budget was added as an int
Third report of one arithmetic mistake, so this time the feature was swept rather than the
line patched. Every seconds-to-milliseconds and budget-plus-backstop site was checked:

  - CN1IntentService.backstopSeconds added the margin as an int and returned an int. With a
    budget the build accepts this went negative, so the service stopped waiting at once and
    tore down the runtime with the handler still in it. Widened, and the Latch it feeds
    takes a long now.
  - AndroidIntentBridge's parked-completion deadline added its margin the same way.
  - The remaining sites are already safe and stay as they are: Intents multiplies by 1000L,
    which promotes the int operand, and the one addition of two compile-time constants
    cannot overflow.

Checked on the compiled port class rather than by reading it. With a budget of
Integer.MAX_VALUE - 1 the backstop is now 2147483651; before the widening the same call
returned -2147483645, which is an await that returns immediately.

+++ Ports/Android/src/com/codename1/impl/android/intents/AndroidIntentBridge.java
@@ -215,7 +215,9 @@ public class AndroidIntentBridge implements IntentBridge {
                 return;
             }
             deadline = System.currentTimeMillis()
-                    + (parkedBudgetSeconds + marginSeconds) * 1000L;
+                    // Widened before the addition: two ints whose sum can exceed what an int
+                    // holds, and a negative deadline makes this return immediately.
+                    + ((long) parkedBudgetSeconds + marginSeconds) * 1000L;
             while (!parkedFinished) {
                 long remaining = deadline - System.currentTimeMillis();
                 if (remaining <= 0) {
